# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SETU-DRR API",
    "description": "Hazard Red Zone Identification, Carrying Capacity Assessment & Relocation Decision Support Platform.",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/hazard/layers": {
      "get": {
        "tags": [
          "Hazard Layers"
        ],
        "summary": "List published static hazard layers",
        "description": "Enumerates every `hazard_static` layer with its cell count, susceptibility range and confidence ceiling, so the client can build a layer switcher without guessing.",
        "operationId": "list_hazard_layers_hazard_layers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HazardLayerSummaryDTO"
                  },
                  "type": "array",
                  "title": "Response List Hazard Layers Hazard Layers Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/hazard/cells": {
      "get": {
        "tags": [
          "Hazard Layers"
        ],
        "summary": "Query a static hazard layer for the map viewport",
        "description": "Returns H3 cells with susceptibility, confidence and coverage provenance, plus the quantile class breaks and confidence ceiling required to render them faithfully.\n\nGeometry is omitted by design \u2014 the client derives hexagon boundaries from the H3 index (deck.gl `H3HexagonLayer`), which is roughly 7x smaller on the wire.",
        "operationId": "get_hazard_cells_hazard_cells_get",
        "parameters": [
          {
            "name": "hazard_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Hazard layer to serve.",
              "examples": [
                "riverine_flood"
              ],
              "default": "riverine_flood",
              "title": "Hazard Type"
            },
            "description": "Hazard layer to serve."
          },
          {
            "name": "res",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 6,
              "description": "H3 grid resolution.",
              "default": 8,
              "title": "Res"
            },
            "description": "H3 grid resolution."
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Viewport as 'min_lon,min_lat,max_lon,max_lat'. Omit for the full layer.",
              "examples": [
                "90.70,26.05,91.45,26.75"
              ],
              "title": "Bbox"
            },
            "description": "Viewport as 'min_lon,min_lat,max_lon,max_lat'. Omit for the full layer."
          },
          {
            "name": "admin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by admin_boundary id or LGD code (e.g. 277 for Barpeta).",
              "title": "Admin"
            },
            "description": "Filter by admin_boundary id or LGD code (e.g. 277 for Barpeta)."
          },
          {
            "name": "min_susceptibility",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Optional susceptibility floor. Left at 0.0 the response still includes hard-zero and no-coverage cells, which the map must distinguish.",
              "default": 0.0,
              "title": "Min Susceptibility"
            },
            "description": "Optional susceptibility floor. Left at 0.0 the response still includes hard-zero and no-coverage cells, which the map must distinguish."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30000,
              "minimum": 1,
              "description": "Maximum cells to return.",
              "default": 20000,
              "title": "Limit"
            },
            "description": "Maximum cells to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardLayerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/hazard/cells/{h3}": {
      "get": {
        "tags": [
          "Hazard Layers"
        ],
        "summary": "Get one cell's hazard dossier and physical drivers",
        "description": "Returns the cell's score with its inundation frequency, HAND, slope and cropland drivers from `hazard_static_flood`, plus confidence already normalised against the layer ceiling.",
        "operationId": "get_hazard_cell_detail_hazard_cells__h3__get",
        "parameters": [
          {
            "name": "h3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "H3 index as hexadecimal string.",
              "examples": [
                "883ce00201fffff"
              ],
              "title": "H3"
            },
            "description": "H3 index as hexadecimal string."
          },
          {
            "name": "hazard_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Hazard layer to read.",
              "default": "riverine_flood",
              "title": "Hazard Type"
            },
            "description": "Hazard layer to read."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HazardCellDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/hazard/summary": {
      "get": {
        "tags": [
          "Hazard Layers"
        ],
        "summary": "Get district-level hazard rollup and officer decision prompt",
        "description": "Returns district-level susceptibility band distributions, habitations & citizens at risk, physical drivers, and an objective factual decision prompt. Districts without computed models (such as Wayanad or Kodagu) return `model_status: 'not_computed'`.",
        "operationId": "get_district_hazard_summary_hazard_summary_get",
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Admin boundary ID or LGD code (e.g. 277 for Barpeta, 98 for Dholpur).",
              "title": "Admin"
            },
            "description": "Admin boundary ID or LGD code (e.g. 277 for Barpeta, 98 for Dholpur)."
          },
          {
            "name": "hazard_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Hazard layer to summarize.",
              "default": "riverine_flood",
              "title": "Hazard Type"
            },
            "description": "Hazard layer to summarize."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictHazardSummaryDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverageFlag": {
        "type": "string",
        "enum": [
          "full",
          "low_coverage",
          "no_coverage"
        ],
        "title": "CoverageFlag",
        "description": "Zonal coverage class for an aggregated H3 cell (Step 10 \u00a710.3).\n\nCRITICAL RENDERING RULE:\nA NO_COVERAGE cell carries susceptibility 0.0 because `apply_quality_flags()` fills\nNaN with zero \u2014 not because it was measured as safe. It must never be drawn with the\nsame treatment as a genuine FR-3.17 hard-zero cell."
      },
...
      "DistrictHazardSummaryDTO": {
        "properties": {
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          },
          "admin_name": {
            "type": "string",
            "title": "Admin Name"
          },
          "lgd_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lgd Code"
          },
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type",
            "default": "riverine_flood"
          },
          "model_status": {
            "type": "string",
            "title": "Model Status",
            "description": "'computed' for SAR-modeled districts; 'not_computed' for unmodeled districts.",
            "default": "computed"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "total_cells": {
            "type": "integer",
            "title": "Total Cells",
            "default": 0
          },
          "coverage": {
            "$ref": "#/components/schemas/HazardLayerCoverageDTO"
          },
          "unmeasured_cells_count": {
            "type": "integer",
            "title": "Unmeasured Cells Count",
            "description": "Number of no_coverage cells. Treated strictly as unmeasured, never safe.",
            "default": 0
          },
          "band_distribution": {
            "$ref": "#/components/schemas/SusceptibilityBandBreakdown"
          },
          "mean_susceptibility": {
            "type": "number",
            "title": "Mean Susceptibility",
            "default": 0.0
          },
          "max_susceptibility": {
            "type": "number",
            "title": "Max Susceptibility",
            "default": 0.0
          },
          "habitations_at_risk_count": {
            "type": "integer",
            "title": "Habitations At Risk Count",
            "default": 0
          },
          "population_at_risk_sum": {
            "type": "integer",
            "title": "Population At Risk Sum",
            "default": 0
          },
          "drivers_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FloodDriverDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_recorded_flood_loss": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Recorded Flood Loss"
          },
          "officer_decision_prompt": {
            "type": "string",
            "title": "Officer Decision Prompt",
            "default": ""
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "admin_id",
          "admin_name"
        ],
        "title": "DistrictHazardSummaryDTO",
        "description": "District-level rollup for officer decision support (GET /hazard/summary)."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Standardized machine-readable error code string."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request tracing identifier."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual error details or validation breakdown."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standardized error details conforming to SETU-DRR API error envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "Error details payload."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Canonical API error response envelope."
      },
...
      "HazardCellDetailDTO": {
        "properties": {
          "h3": {
            "type": "string",
            "title": "H3"
          },
          "h3_int": {
            "type": "integer",
            "title": "H3 Int"
          },
          "res": {
            "type": "integer",
            "title": "Res"
          },
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "susceptibility": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Susceptibility"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "confidence_normalised": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence Normalised",
            "description": "confidence / layer ceiling \u2014 the value safe to display."
          },
          "quality_flag": {
            "$ref": "#/components/schemas/CoverageFlag"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid",
            "description": "[longitude, latitude]"
          },
          "admin_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Name"
          },
          "population": {
            "type": "number",
            "title": "Population",
            "default": 0.0
          },
          "is_permanent_red_candidate": {
            "type": "boolean",
            "title": "Is Permanent Red Candidate",
            "description": "susceptibility >= PRZ_ANY_SUSCEPTIBILITY (FR-3.9).",
            "default": false
          },
          "drivers": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FloodDriverDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "h3",
          "h3_int",
          "res",
          "hazard_type",
          "susceptibility",
          "confidence",
          "confidence_normalised",
          "quality_flag",
          "model_version",
          "centroid"
        ],
        "title": "HazardCellDetailDTO",
        "description": "Full per-cell dossier payload (GET /hazard/cells/{h3})."
      },
...
      "HazardLayerCoverageDTO": {
        "properties": {
          "full": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Full",
            "default": 0
          },
          "low_coverage": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Low Coverage",
            "default": 0
          },
          "no_coverage": {
            "type": "integer",
            "minimum": 0.0,
            "title": "No Coverage",
            "default": 0
          }
        },
        "type": "object",
        "title": "HazardLayerCoverageDTO",
        "description": "Population counts per coverage class, for the legend and the empty-state copy."
      },
      "HazardLayerLegendDTO": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Classification method used to derive breaks.",
            "default": "quantile"
          },
          "quantiles": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Quantiles",
            "description": "Quantile positions the breaks were sampled at."
          },
          "breaks": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Breaks",
            "description": "Ascending class break values in susceptibility units."
          },
          "domain": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Domain",
            "description": "[min, max] observed susceptibility across the layer."
          },
          "confidence_ceiling": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Confidence Ceiling",
            "description": "Maximum confidence present in the layer. Divide each cell's confidence by this to obtain a displayable [0, 1] value."
          },
          "prz_susceptibility_threshold": {
            "type": "number",
            "title": "Prz Susceptibility Threshold",
            "description": "FR-3.9 Permanent Red Zone susceptibility cut (core.constants.PRZ_ANY_SUSCEPTIBILITY)."
          }
        },
        "type": "object",
        "required": [
          "quantiles",
          "breaks",
          "domain",
          "confidence_ceiling",
          "prz_susceptibility_threshold"
        ],
        "title": "HazardLayerLegendDTO",
        "description": "Quantile class breaks and normalisation ceilings for the active layer.\n\nComputed server-side over the queried population rather than assumed client-side.\nA linear 0->1 ramp renders the Barpeta flood layer almost uniformly: half its cells\nfall between 0.39 and 0.48 because mean HAND is 1.83 m across the floodplain, which\nsaturates the `1 - HAND/P99` term. The signal lives in the top decile."
      },
      "HazardLayerResponse": {
        "properties": {
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type",
            "description": "Hazard type served, e.g. 'riverine_flood'."
          },
          "res": {
            "type": "integer",
            "title": "Res",
            "description": "H3 resolution of the returned cells."
          },
          "count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Count",
            "description": "Number of cells in this response."
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "description": "True when the limit clipped the result set.",
            "default": false
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "Pipeline model version tag.",
            "default": "v1.0.0"
          },
          "legend": {
            "$ref": "#/components/schemas/HazardLayerLegendDTO"
          },
          "coverage": {
            "$ref": "#/components/schemas/HazardLayerCoverageDTO"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/HazardCellDTO"
            },
            "type": "array",
            "title": "Cells"
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "hazard_type",
          "res",
          "count",
          "legend",
          "coverage"
        ],
        "title": "HazardLayerResponse",
        "description": "Envelope for GET /hazard/cells \u2014 cells plus everything needed to colour them."
      },
      "HazardLayerSummaryDTO": {
        "properties": {
          "hazard_type": {
            "type": "string",
            "title": "Hazard Type"
          },
          "res": {
            "type": "integer",
            "title": "Res"
          },
          "cell_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cell Count"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "min_susceptibility": {
            "type": "number",
            "title": "Min Susceptibility"
          },
          "max_susceptibility": {
            "type": "number",
            "title": "Max Susceptibility"
          },
          "mean_susceptibility": {
            "type": "number",
            "title": "Mean Susceptibility"
          },
          "confidence_ceiling": {
            "type": "number",
            "title": "Confidence Ceiling"
          }
        },
        "type": "object",
        "required": [
          "hazard_type",
          "res",
          "cell_count",
          "model_version",
          "min_susceptibility",
          "max_susceptibility",
          "mean_susceptibility",
          "confidence_ceiling"
        ],
        "title": "HazardLayerSummaryDTO",
        "description": "One available static hazard layer (GET /hazard/layers)."
      },
...
    }
  }
}